+++ src/idc/filter/objdet/__init__.py
@@ -1,3 +1,4 @@
+from ._check_polygon_validity import CheckPolygonValidity, ACTIONS, ACTION_RECORD_ALL, ACTION_RECORD_INVALID, ACTION_RECORD_VALID
 from ._coerce_box import CoerceBox
 from ._coerce_mask import CoerceMask
 from ._dimension_discarder import DimensionDiscarder