+++ CHANGES.rst
@@ -46,6 +46,8 @@ Changelog
 - introduced the `IDC_EXIF_AUTOROTATE` environment variable to implicitly apply the EXIF auto-rotation when loading an image
 - `idc-exec` can specify global options to the conversion via `--exec_global_opts`
 - the writers `to-adams-ic` and `to-adams-od` now allow specifying the file extension to use for the report files
+- added `check-polygon-validity` object detection filter that can check for, e.g., self-intersecting polygons
+  using shapely's `shapely.validation.explain_validity` function.
 
 
 0.1.0 (2025-10-31)